ColladaLoader: Split into modular files
Refactored ColladaLoader from a single 4000+ line file into three modules:

- ColladaLoader.js - Entry point
- ColladaParser.js - XML parsing (parse* functions)
- ColladaComposer.js - Three.js object building (build*/get* functions)
